requirements.txt
config
kafka-python